refactor(#87): isAuthUser를 feature/user/utils로 분류하기
