+++ apps/frontend/src/page/(main)/lotus/create/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute, redirect } from '@tanstack/react-router';
-import { isAuthUser } from '@/shared/utils/isAuthUser';
+import { isAuthUser } from '@/feature/User/utils/isAuthUser';
 
 export const Route = createFileRoute('/(main)/lotus/create/')({
   beforeLoad: () => {
