+++ apps/frontend/src/page/(main)/user/index.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute, redirect } from '@tanstack/react-router';
 import { z } from 'zod';
-import { isAuthUser } from '@/shared/utils/isAuthUser';
+import { isAuthUser } from '@/feature/User/utils/isAuthUser';
 
 const userLotusSearchValidation = z.object({
   page: z.number().default(1)
