--- apps/frontend/src/shared/utils/isAuthUser.ts
@@ -1,8 +1,0 @@
-import { queryClient } from '@/app/query';
-import { UserType } from '@/feature/User/type';
-
-export const isAuthUser = () => {
-  const user = queryClient.getQueryData<UserType>(['user']);
-
-  return !!user;
-};
